Add files via upload
I have completed the SQL challenge conducted by Digits n Data. In this challenge I have used Mysql to solve the question. we have Shows, Partnerships, ShowPartnerships, Awards and OnlinePresence tables present in this challenge

📻 92.7 Big FM's history
challenge is to craft SQL queries to extract insights from the 92.7 BIG FM radio network database. We have multiple table in the dataset from this we have to find the solution for our quries.

Questions:
1)Retrieve all stations in the "East" region?
2)List all shows hosted by "Vrajesh Hirjee"?
3)Count the number of awards each show has won?
4)Find shows that have partnerships with "Spotify"?
5)Retrieve hosts who joined before 2010?
6)List the shows and their launch dates in descending order of launch date?
7)Find the total count of shows for each host?
8)Show the online presence platforms with their links?
9)Retrieve the stations in the "South" region launched before 2010?
10)Rank hosts based on the number of shows they have hosted?